plot msv: correctly add datasets if sampling intervals differ
The total energy of a system with Nosé–Hoover chain consists of the total
energy of the particles and the total energy of the chain. Both data sets may
be stored at different points in time.

The issue can neatly be solved by apply NumPy's set operations on the 'step'
datasets.